splitting up baggers
